pyproject.toml
[project]
name = "txt-to-table-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain",
    "langchain-community>=0.4.1",
    "langchain_google_genai",
    "pydantic[email]>=2.12.5",
    "pypdf>=6.9.0",
    "python-dotenv>=1.2.2",
]
